+++ lib/utils/characters.dart
@@ -1,0 +1,7 @@
+class Chars {
+  const Chars._();
+  static const lower = 'abcdefghijklmnopqrstuvwxyz';
+  static const upper = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
+  static const number = '0123456789';
+  static const special = '!@#%^&*()_+-=[]{};:,.<>/?';
+}