+++ lib/utils/dialog_helper.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 /// actions: a builder with dismiss callback to dismiss the dialog
 Future<T?> showAlertDialog<T>(
   BuildContext context, {
-  Widget? title,
-  Widget? content,
+  String? title,
+  String? content,
   List<Widget>? Function(VoidCallback dismiss)? actions,
   barrierDismissible = true,
 }) {
@@ -13,8 +13,8 @@ Future<T?> showAlertDialog<T>(
     barrierDismissible: barrierDismissible,
     builder: (context) {
       return AlertDialog(
-        title: title,
-        content: content,
+        title: title != null ? Text(title) : null,
+        content: content != null ? Text(content) : null,
         actions: actions?.call(Navigator.of(context).pop),
       );
     },