+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.4
+* fix chars
+
+## 1.0.3
+* Add Chars, copy, toast
+
+## 1.0.2
+* Add random in range
+
 ## 1.0.1
 * Add dialog helper